- ordered configuration same as in OCP - deleted timer keys - deleted timeout modal component - deleted session timer mixin - deleted timed module - deleted timer state properties - renamed models -> interfaces - deleted coundown timer component - deleted time state, actions, mutations - deleted unused Analyze Name module
