+++ src/components/mixins/displayed-component-mixin.ts
@@ -2,7 +2,6 @@ import Vue from 'vue'
 import { Component } from 'vue-property-decorator'
 
 import newRequestModule from '@/store/new-request-module'
-import timerModule from '@/modules/vx-timer'
 
 @Component
 export default class DisplayedComponentMixin extends Vue {
@@ -25,7 +24,6 @@ export default class DisplayedComponentMixin extends Vue {
     // Are we cancelling a new NR?
     if (['SubmissionTabs'].indexOf(componentName) > -1) {
       await newRequestModule.cancelAnalyzeName('Tabs')
-      timerModule.stopTimer({ id: this.$NR_COMPLETION_TIMER_NAME })
       this.redirectToStart()
     } else if (['ExistingRequestDisplay'].indexOf(componentName) > -1) {
       await newRequestModule.cancelAnalyzeName('Tabs')
@@ -35,7 +33,6 @@ export default class DisplayedComponentMixin extends Vue {
       // Check in the NR to release the INPROGRESS lock on the NR
       await newRequestModule.cancelEditExistingRequest()
       await newRequestModule.checkinNameRequest()
-      timerModule.stopTimer({ id: this.$EXISTING_NR_TIMER_NAME })
       this.redirectToStart()
     } else {
       await newRequestModule.cancelAnalyzeName('Tabs')
