+++ src/components/payment/payment-complete-modal.vue
@@ -78,7 +78,6 @@ export default class PaymentCompleteModal extends Mixins(
     const { sessionPaymentId, sessionPaymentAction } = this
     // Check for a payment ID in sessionStorage, if it has been set, we've been redirected away from the application,
     // and need to rehydrate the application using the payment ID (for now, it could be some other token too)!
-    // TODO: Set the timer here!
     if (sessionPaymentId && sessionPaymentAction) {
       // Make sure edit mode is disabled or it will screw up the back button
       await newRequestModule.mutateEditMode(false)
