+++ src/interfaces/business.ts
