+++ src/modules/payment/store/getters.ts
@@ -42,7 +42,6 @@ export default {
     }
   },
   [types.GET_PAYMENT_REQUEST]: state => {
-    // TODO: Zap the payment request when timer expires?
     let paymentRequestState
     if ((!state[STATE_KEY].request || Object.keys(state[STATE_KEY].request).length === 0) &&
       sessionStorage.getItem('paymentRequest')) {
