+++ src/modules/payment/store/types.ts
@@ -41,5 +41,3 @@ export const SET_SBC_PAYMENT = 'setSbcPayment'
 export const SET_PAYMENT = 'setPayment'
 export const SET_PAYMENT_RECEIPT = 'setPaymentReceipt'
 export const SET_PAYMENTS = 'setPayments'
-export const START_PAYMENT_TIMEOUT_TIMER = 'startPaymentTimeoutTimer'
-export const CLEAR_PAYMENT_TIMEOUT_TIMER = 'clearPaymentTimeoutTimer'
