--- src/modules/vx-timer/store/types.ts
@@ -1,6 +1,0 @@
-export const GET_TIMER = 'getTimer'
-export const CREATE_AND_START_TIMER = 'createAndStartTimer'
-// export const CREATE_PROMISE_TIMER = 'createPromiseTimer'
-export const RESTART_TIMER = 'restartTimer'
-export const REFRESH_TIMER = 'refreshTimer'
-export const STOP_TIMER = 'stopTimer'
