+++ src/plugins/getConfig.ts
@@ -1,14 +1,7 @@
 import axios from 'axios'
 
 export interface EnvConfigI {
-    $APP_IS_RUNNING_LOCALLY?: boolean
     $PAYMENT_PORTAL_URL?: string
-    // NEW NR completion timer
-    $NR_COMPLETION_TIMER_NAME?: string
-    $NR_COMPLETION_TIMEOUT_MS?: number
-    // EXISTING NR timer
-    $EXISTING_NR_TIMER_NAME?: string
-    $EXISTING_NR_TIMEOUT_MS?: number
 }
 
 // Detailed commentary on how this works @ github.com/bcgov/business-filings-ui/utils/fetch-config.ts
@@ -28,13 +21,6 @@ export function getConfig (): Promise<EnvConfigI> {
     const baseURL = response.data[0]['URL'] + '/api/v1'
 
     const paymentUrl = response.data[0]['PAYMENT_PORTAL_URL']
-    const nrCompletionTimerName = response.data[0]['NR_COMPLETION_TIMER_NAME']
-    const nrCompletionTimeoutMs = response.data[0]['NR_COMPLETION_TIMEOUT_MS']
-    const nrCompletionTimeoutMs_dev = response.data[0]['DEV_NR_COMPLETION_TIMEOUT_MS']
-
-    const existingNrTimerName = response.data[0]['EXISTING_NR_TIMER_NAME']
-    const existingNrTimeoutMs = response.data[0]['EXISTING_NR_TIMEOUT_MS']
-    const existingNrTimeoutMs_dev = response.data[0]['DEV_EXISTING_NR_TIMEOUT_MS']
 
     const keycloakConfigPath = response.data[0]['KEYCLOAK_CONFIG_PATH']
     sessionStorage.setItem('KEYCLOAK_CONFIG_PATH', keycloakConfigPath)
@@ -86,16 +72,8 @@ export function getConfig (): Promise<EnvConfigI> {
 
     axios.defaults.baseURL = baseURL
 
-    const appIsRunningLocally = !!(window['webpackHotUpdate'])
-
     const config: EnvConfigI = {
-      // NB: window['webpackHotUpdate'] is a function when running locally, otherwise it's undefined
-      $APP_IS_RUNNING_LOCALLY: appIsRunningLocally,
-      $PAYMENT_PORTAL_URL: paymentUrl,
-      $NR_COMPLETION_TIMER_NAME: nrCompletionTimerName,
-      $NR_COMPLETION_TIMEOUT_MS: appIsRunningLocally ? nrCompletionTimeoutMs_dev : nrCompletionTimeoutMs,
-      $EXISTING_NR_TIMER_NAME: existingNrTimerName,
-      $EXISTING_NR_TIMEOUT_MS: appIsRunningLocally ? existingNrTimeoutMs_dev : existingNrTimeoutMs
+      $PAYMENT_PORTAL_URL: paymentUrl
     }
     resolve(config)
   })
