--- src/store/getters/index.ts
@@ -1,1 +1,0 @@
-export * from './getters'
