--- src/store/types.ts
@@ -1,4 +1,0 @@
-export const SHOW_NR_SESSION_EXPIRY_MODAL = 'showNrSessionExpiryModal'
-export const HIDE_NR_SESSION_EXPIRY_MODAL = 'hideNrSessionExpiryModal'
-// export const SET_ROLLBACK_ON_EXPIRE = 'setRollbackOnExpire' // NOT USED
-// export const SET_CHECK_IN_ON_EXPIRE = 'setCheckInOnExpire' // NOT USED
